unit tests fix
